% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_work.R
\name{calc_work}
\alias{calc_work}
\title{Calculate work from slope raster and body size}
\usage{
calc_work(
  slope = NULL,
  m = NULL,
  output_to_disk = FALSE,
  output_file = NULL,
  g = 9.80665,
  J_to_Kcal = 4184,
  work_in_kcal = FALSE
)
}
\arguments{
\item{slope}{is the slope raster}

\item{m}{is the body mass (kg) of the species}

\item{output_to_disk}{if to write output to disk}

\item{output_file}{is the output destiation, if @param output_to_disc == TRUE}

\item{g}{is the gravity force}

\item{J_to_Kcal}{is the Joule to kilocalory conversion contant}

\item{work_in_kcal}{if the output should be expressed in kcal}
}
\value{
a raster of the energetic costs of locomotion
}
\description{
Calculate work from slope raster and body size
}
