% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuitscape.R
\name{circuitscape_skeleton}
\alias{circuitscape_skeleton}
\title{Create the initialization file for the julia package Circuitscape}
\usage{
circuitscape_skeleton(en = NULL, path = NULL, points = NULL)
}
\arguments{
\item{en}{an enerscape object.}

\item{path}{full path where to write the .ini file.}

\item{points}{data.frame with origin and destination coordinates.}
}
\value{
Nothing, only write the circuitscape.ini file to disk.
}
\description{
This creates the init file for the julia package Circuitscape:
\url{https://juliapackages.com/p/circuitscape}.
}
