% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dem}
\alias{dem}
\title{Digital elevation model of the Sirente-Velino Regional Park}
\format{
A raster layer with digital elevation data (meters)
}
\source{
\url{http://tinitaly.pi.ingv.it/}
}
\usage{
dem
}
\description{
Digital elevation model of the Sirente-Velino Regional Park
}
\keyword{datasets}
