% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_to_slope.R
\name{dem_to_slope}
\alias{dem_to_slope}
\title{Calculate slope from digital elevation model}
\usage{
dem_to_slope(dem)
}
\arguments{
\item{dem}{raster file of the digital elevation model, either a raster
or a full path location of the file}
}
\value{
raster of slope in degrees
}
\description{
Calculate slope from digital elevation model
}
