% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distances}
\alias{distances}
\title{Spatial distances}
\usage{
distances(x, center, res)
}
\arguments{
\item{x}{matrix with values}

\item{center}{numeric value (double) with the value of the focal cell}

\item{res}{numeric value (double) of the spatial resolution of the matrix}
}
\value{
Vector with values the distances between x and center
}
\description{
Spatial distances
}
