% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enerscape_internals.R
\name{.calc_arc_cond}
\alias{.calc_arc_cond}
\title{Compute conductance}
\usage{
.calc_arc_cond(slope, m, work_in_kcal = TRUE, j_to_kcal = 4184)
}
\arguments{
\item{slope}{slope transition matrix.}

\item{m}{species body mass (kg).}

\item{work_in_kcal}{if work should be expressed in kilocalories.}

\item{j_to_kcal}{joules to kilocalories conversion constant.}
}
\value{
A transition layer with values the conductance between cells, i.e.
  the distance that can be traveled per unit of energy (1 / J or 1 / kcal).
}
\description{
Internal function for enerscape - calculate conductance
}
\details{
Internal function of enerscape, don't call directly.
}
