% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enerscape_internals.R
\name{.calc_cond}
\alias{.calc_cond}
\title{Internal function for enerscape - calculate conductance}
\usage{
.calc_cond(slope, m, work_in_kcal = TRUE, j_to_kcal = 4184)
}
\arguments{
\item{slope}{slope transition matrix.}

\item{m}{species body mass (kg).}

\item{work_in_kcal}{if work should be expressed in kilocalories.}

\item{j_to_kcal}{joules to kilocalories conversion constant.}
}
\description{
Internal function for enerscape - calculate conductance
}
