% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement.R
\name{.calc_work}
\alias{.calc_work}
\title{Internal function for en_shortest_path - calculate work}
\usage{
.calc_work(x, m, J_to_kcal = 4184, work_in_kcal = TRUE)
}
\arguments{
\item{x}{a slope transition layer}

\item{m}{is the body mass (kg) of the species}

\item{J_to_kcal}{is the Joule to kilocalory conversion contant}

\item{work_in_kcal}{if the output should be expressed in kcal}
}
\value{
a raster of the energetic costs of locomotion
}
\description{
Internal function for en_shortest_path - calculate work
}
