% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{.flip}
\alias{.flip}
\title{Flip aspect terrain horizontally, i.e. East to West and West to East}
\usage{
.flip(x)
}
\arguments{
\item{x}{raster of terrain aspect}
}
\description{
Flip aspect terrain horizontally, i.e. East to West and West to East
}
