% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{.rotate}
\alias{.rotate}
\title{Rotate aspect raster to orient it northward}
\usage{
.rotate(x, angle = 90)
}
\arguments{
\item{x}{raster of terrain aspect}

\item{angle}{angle of rotation, 90 for North = 90 degrees}
}
\description{
Rotate aspect raster to orient it northward
}
