% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement.R
\name{en_shortest_path}
\alias{en_shortest_path}
\title{Calculate shortest path between origin and destination}
\usage{
en_shortest_path(
  x,
  m,
  neigh = 4,
  or,
  dest,
  simulate_random_points = FALSE,
  rep = 10
)
}
\arguments{
\item{x}{a digital elevation model raster layer}

\item{m}{species body mass (kg).}

\item{neigh}{number of neighbours cells to consider}

\item{or}{origin point}

\item{dest}{destination point}

\item{simulate_random_points}{if to simulate shortest path among random
points. default = FALSE}

\item{rep}{number or random origin and destination points if
simulate_random_points = TRUE. default = 10}
}
\description{
Calculate shortest path between origin and destination
}
