% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{energyHuman}
\alias{energyHuman}
\title{Energy Landscape for walking people}
\usage{
energyHuman(mass, v, slope, distance, res, kcal)
}
\arguments{
\item{mass}{body mass of species (kg)}

\item{v}{speed}

\item{slope}{vector with slopes}

\item{distance}{vector with distances}

\item{res}{numeric value (double) of the spatial resolution of the matrix}

\item{kcal}{(boolean) if to return the result in kCal (true) or J (false)}
}
\value{
Vector with the energy cost of locomotion (EnergyScape)
}
\description{
Energy Landscape for walking people
}
