% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{energyscapeHuman}
\alias{energyscapeHuman}
\title{Energy Landscape}
\usage{
energyscapeHuman(x, mass, v, n, res, kcal)
}
\arguments{
\item{x}{matrix with values the elevation.}

\item{mass}{body mass of species (kg).}

\item{v}{speed}

\item{n}{(integer) number of neighbours to consider (either 4 or 8).}

\item{res}{numeric value (double) of the spatial resolution of the matrix.}

\item{kcal}{(boolean) if to return the result in kCal (true) or J (false).}
}
\value{
Matrix with the energy cost of locomotion (EnergyScape).
}
\description{
Energy Landscape
}
