% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enerscape.R
\name{enerscape}
\alias{enerscape}
\title{Calculate the energy landscape}
\usage{
enerscape(dem, m, unit = "joule", neigh = 4)
}
\arguments{
\item{dem}{raster file of the digital elevation model, either a raster or a
full path location of the file.}

\item{m}{species body mass (kg).}

\item{unit}{if joules ('joule') or kilocalories ('kcal').}

\item{neigh}{number of neighbors used to calculate slope and work. See
?raster::terrain for details.}
}
\description{
Calculate the energy landscape
}
