% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enerscape_gridded.R
\name{enerscape_gridded}
\alias{enerscape_gridded}
\title{Runs enerscape on blocks of a RasterLayer.}
\usage{
enerscape_gridded(
  dem,
  m,
  split.hor,
  split.vert,
  unit = "joule",
  neigh = 16,
  method = "ARC",
  v = NULL,
  write = FALSE,
  out.dir = raster::tmpDir()
)
}
\arguments{
\item{dem}{raster file of the digital elevation model, either a raster or a
full path location of the file.}

\item{m}{species body mass (kg).}

\item{split.hor}{integer, number of horizontal splits.}

\item{split.vert}{integer, number of vertical splits.
avoid missing values at the polygon border when merging back together.}

\item{unit}{if joules ('joule') or kilocalories ('kcal').}

\item{neigh}{number of neighbor cells that are connected together.}

\item{method}{method to use to compute the energy costs. 'ARC' refers to the
model from Pontzer (2016) and 'cycling' to the model for cyclist from di
Prampero et al. (1979).}

\item{v}{speed of cyclist (km / h), only for \code{method = 'cycling'}.}

\item{write}{TRUE/FALSE, if to write the raster to disk.}

\item{out.dir}{string, output directory; default to tempdir().}
}
\value{
A list with elements a rasterStack of the digital elevation model,
  slope, energy landscape, and conductance and the conductance as a transitionLayer for
  path analysis.
}
\description{
This is the main function to compute energy landscapes from a digital
elevation model and body mass of animals based on the model from Pontzer
(2016). The core of the computations are done using the \emph{gdistance}
(Etten, 2017) package.
}
\details{
Split the DEM into blocks and run enerscape for each block
 sequentially.
This is useful when the DEM layer is particularly large. There are two
options: the first is to manipulate all layers in memory (write = FALSE);
the second is to write each layer to disk and load it when necessary.
The former is faster, but consume more RAM.
}
\examples{
\donttest{
library(raster)
library(enerscape)
data("volcano")
r <- raster(volcano)
crs(r) <- "+proj=longlat +datum=WGS84 +no_defs"
r <- projectRaster(r, crs = "EPSG:32759", res = c(1e3, 1e3))
en <- enerscape(r, 1)$rasters$EnergyScape
en_gr <- enerscape_gridded(r, 1, split.hor = 3, split.vert = 3)
}
}
\references{
Etten, J. van. (2017). R Package gdistance: Distances and Routes
  on Geographical Grids. Journal of Statistical Software, 76(1), 1–21.
  \doi{https://doi.org/10.18637/jss.v076.i13}.

  Pontzer, H. (2016). A unified theory for the energy cost of legged
  locomotion. Biology Letters, 12(2), 20150935. \doi{
  https://doi.org/10.1098/rsbl.2015.0935}.

  di Prampero, P. E., Cortili, G., Mognoni, P., & Saibene, F. (1979).
  Equation of motion of a cyclist. Journal of Applied Physiology, 47(1),
  201–206. \doi{https://doi.org/10.1152/jappl.1979.47.1.201}
}
