% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enerscape_gridded.R
\name{enerscape_merge}
\alias{enerscape_merge}
\title{Merge together EnergyScapes for a gridded enerscape}
\usage{
enerscape_merge(writedir = raster::tmpDir(), grid, dem)
}
\arguments{
\item{writedir}{directory to write EnergyScapes for each grid; default to tmpDir().}

\item{grid}{enerscape grid, as created using the \code{make_grid()} function.}

\item{dem}{raster file of the digital elevation model, either a raster or a
full path location of the file.}
}
\value{
rasterLayer of the merged and corrected EnergyScapes.
}
\description{
This merge back together EnergyScapes computed using \code{enerscape_gridded()}
correcting for artifacts generated by the splitting.
}
\details{
\code{enerscape_gridded()} runs enerscape for each block of the grid
 sequentially and writes (only) the rasterLayer of the EnergyScape to the specified
 folder, default to the raster temporary directory. This temporary files are
 then merge back together by \code{enerscape_merge()}. This function also calculates
 enerscape for the vertical and horizontal strips used to split \code{dem} into blocks,
 which are then used to correct for artifacts of the splitting.
}
\examples{
\dontrun{
library(raster)
library(enerscape)
dem <- getData("alt", path = tmpDir(), country = "austria")
dem <- projectRaster(dem, crs = crs("EPSG:31255"), res = c(1000, 1000))
en_grid <- make_grid(extent(dem), crs(dem), 10, 10)
enerscape_gridded(dem, 10, grid = en_grid)
en <- enerscape_merge(tmpDir(), en_grid, dem)
}
}
\references{
Etten, J. van. (2017). R Package gdistance: Distances and Routes
  on Geographical Grids. Journal of Statistical Software, 76(1), 1–21.
  \doi{https://doi.org/10.18637/jss.v076.i13}.

  Pontzer, H. (2016). A unified theory for the energy cost of legged
  locomotion. Biology Letters, 12(2), 20150935. \doi{
  https://doi.org/10.1098/rsbl.2015.0935}.

  di Prampero, P. E., Cortili, G., Mognoni, P., & Saibene, F. (1979).
  Equation of motion of a cyclist. Journal of Applied Physiology, 47(1),
  201–206. \doi{https://doi.org/10.1152/jappl.1979.47.1.201}
}
