% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{Calculate the gradient relative to the minimum distance between two points}
\usage{
gradient(dem, origin, destination)
}
\arguments{
\item{dem}{digital elevation model raster}

\item{origin}{data.frame with x and y coordinates of origin point}

\item{destination}{data.frame with x and y coordinates of destination point}
}
\description{
Calculate the gradient relative to the minimum distance between two points
}
\details{
This function return the gradient of the slope evaluated considering
  movement always parallel to the minimum distance axis between the origin
  and destination points
}
