% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enerscape_gridded.R
\name{gridPolygon}
\alias{gridPolygon}
\title{Create a gridPolygon}
\usage{
gridPolygon(mat, proj4, ID)
}
\arguments{
\item{mat}{matrix with coordinates.}

\item{proj4}{string, proj4string of the original raster.}

\item{ID}{integer, ID of the polygon (can be anything as lons as unique).}
}
\value{
a SpatialPolygonsDataFrame.
}
\description{
Create a gridPolygon
}
\details{
gridPolygon() is used within gridRaster() and is mostly for enerscape
package internal functioning.
}
\examples{
mat <- matrix(c(0, 0, 1, 0, 1, 1, 0, 1, 0, 0), ncol = 2, byrow = TRUE)
proj4 <- "+proj=longlat +datum=WGS84 +no_defs"
gridPolygon(mat, proj4, ID = 1)
}
