% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enerscape_gridded.R
\name{gridRaster}
\alias{gridRaster}
\title{Create spatial grid}
\usage{
gridRaster(r, split.hor, split.vert)
}
\arguments{
\item{r}{RasterLayer.}

\item{split.hor}{integer, number of horizontal splits.}

\item{split.vert}{integer, number of vertical splits.}
}
\value{
a SpatialPolyonsDataFrame.
}
\description{
Create spatial grid
}
\details{
gridRaster makes a spatial grid that will be used to divide
the raster into smaller units for computations. The final grid may not
necessarily be split into split.hor times split.vert blocks, but it can happen
that the final grid has (split.hor - 1) times (split.vert - 1) blocks.
}
\examples{
\donttest{
library(raster)
data("volcano")
r <- raster(volcano)
crs(r) <- "+proj=longlat +datum=WGS84 +no_defs"
gridRaster(r, split.hor = 3, split.vert = 3)
}
}
