% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enerscape_gridded.R
\name{make_grid}
\alias{make_grid}
\title{Create spatial grid}
\usage{
make_grid(ext, crs, horizontal_grids, vertical_grids)
}
\arguments{
\item{ext}{extent of x, x@extent}

\item{crs}{crs of x, x@crs}

\item{horizontal_grids}{integer, how many horizontal lines should the grid have.}

\item{vertical_grids}{integer, how many vertical lines should the grid have.}
}
\value{
A named list with the grid (\code{grid}), the xy coordinates
 of its node points (\code{grid_xy}), horizontal (\code{horizontals}) and
 vertical (\code{verticals}) spatial SpatialLines, and adjustment value
 to get correct grid assignment (\code{adjust}). In most cases, you can ignore
 this list and just pass it to \code{enerscape_gridded()} and
 \code{enerscape_merge()} function as it is.
}
\description{
This creates a spatial grid from a rasterLayer (usually the DEM).
This grid is then used to run \code{enerscape()} on each grid
separately.
}
\examples{
\dontrun{
library(raster)
library(enerscape)
dem <- getData("alt", path = tmpDir(), country = "austria")
dem <- projectRaster(dem, crs = crs("EPSG:31255"), res = c(1000, 1000))
enerscape_gridded(dem, 10, grid = en_grid)
en <- enerscape_merge(tmpDir(), en_grid, dem)
}
}
