% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enerscape_gridded.R
\name{mergeRaster}
\alias{mergeRaster}
\title{Merge split raster blocks}
\usage{
mergeRaster(x = NULL, out.dir = NULL, pattern = "")
}
\arguments{
\item{x}{an enerscape rasterBlocks class, which is a list of raster blocks.}

\item{out.dir}{string, output directory; default to tempdir().}

\item{pattern}{string, regular pattern to subset files in out.dir.}
}
\value{
a RasterLayer.
}
\description{
Merge split raster blocks
}
\details{
if x == NULL, rasterMerge() loads the rasters save to disk by
rasterSplit()
}
\examples{
\donttest{
library(raster)
data("volcano")
r <- raster(volcano)
crs(r) <- "+proj=longlat +datum=WGS84 +no_defs"
gr <- gridRaster(r, split.hor = 3, split.vert = 3)
splits <- splitRaster(r, gr)
}
}
