% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{neighbours}
\alias{neighbours}
\title{Neighbours}
\usage{
neighbours(i, j, n, x, direction)
}
\arguments{
\item{i}{row index.}

\item{j}{column index.}

\item{n}{number of neighbours (4 or 8).}

\item{x}{matrix with values.}

\item{direction}{(integer) for which direction to calculate costs: 0 for all, 1 for left, 2 for down, 3 for right, 4 for up.}
}
\value{
Vector with values the neighours of x
}
\description{
Neighbours
}
