% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enerscape.R
\name{plot_enerscape}
\alias{plot_enerscape}
\title{Plot enerscape object}
\usage{
plot_enerscape(
  en,
  what = "all",
  contour = TRUE,
  n_contour = 10,
  axes = FALSE,
  max_quantile = 1
)
}
\arguments{
\item{en}{is the output of the enerscape() function}

\item{what}{specifies what to plot. Available are "work", "slope", or "all"}

\item{contour}{specifies if DEM contour plots should be overlayed on the plot}

\item{n_contour}{specifies how many levels of DEM contours should be plotted,
if any}

\item{axes}{specifies if axes should be added to the plot}

\item{max_quantile}{specifies the maximum quantile to be displayed, i.e.
values above this quantile are re-assigned to the maximum value within the
quantile. This is useful for plotting, as sometimes few cells have
extremely high values, due to almost vertical displacement. In such cases,
a max_quantile = 0.99 removes these outliers.}
}
\description{
Plot enerscape object
}
