% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pontzer}
\alias{pontzer}
\title{Energy cost of transport from Pontzer (2016)}
\format{
A data frame with 92 rows and 5 variables:
\describe{
  \item{Species}{species name}
  \item{Incline}{incline of movement}
  \item{Mass}{species body mass}
  \item{Cost.of.Transport}{cost of transport}
  \item{Source}{original source of data}
}
}
\source{
\doi{https://doi.org/10.1098/rsbl.2015.0935}
}
\usage{
pontzer
}
\description{
Energy cost of transport from Pontzer (2016)
}
\references{
#' Pontzer, H. (2016). A unified theory for the energy cost of legged
locomotion. Biology Letters, 12(2), 20150935.
}
\keyword{datasets}
