% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sirente}
\alias{sirente}
\title{Monte Sirente Digital Elevation Model}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 393 rows and 594 columns.
}
\usage{
sirente
}
\description{
A matrix with the digital elevation mode of the Monte Sirente (Central Italy).
}
\keyword{datasets}
