% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enerscape_gridded.R
\name{splitRaster}
\alias{splitRaster}
\title{Split raster into blocks}
\usage{
splitRaster(x, y, buffer = TRUE, write = FALSE, out.dir = tempdir())
}
\arguments{
\item{x}{RasterLayer.}

\item{y}{SpatialPolygonsDataFrame, polygon grid (as created using
gridRaster()),}

\item{buffer}{TRUE/FALSE, if to create a small buffer around the polygon to
avoid missing values at the polygon border when merging back together.}

\item{write}{TRUE/FALSE, if to write the raster to disk.}

\item{out.dir}{string, output directory; default to tempdir().}
}
\value{
an enerscape rasterBlocks objects, which is a list of RasterLayers.
}
\description{
Split raster into blocks
}
\details{
as splitting exactly at the spatialPolygons boundaries will create
later artifacts at the polygons edge, it is better to always set
buffer = TRUE, which will polygons buffered by twice the
resolution of the raster to split before cropping.
}
\examples{
\donttest{
library(raster)
data("volcano")
r <- raster(volcano)
crs(r) <- "+proj=longlat +datum=WGS84 +no_defs"
gr <- gridRaster(r, split.hor = 3, split.vert = 3)
splits <- splitRaster(r, gr)
}
}
