% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement.R
\name{.calc_slope}
\alias{.calc_slope}
\title{Internal function for en_shortest_path - calculate slope}
\usage{
.calc_slope(x)
}
\arguments{
\item{x}{a digital elevation model raster layer}
}
\description{
Internal function for en_shortest_path - calculate slope
}
