% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enerscape_internals.R
\name{.calc_slope}
\alias{.calc_slope}
\title{Internal function for enerscape - calculate slope}
\usage{
.calc_slope(x)
}
\arguments{
\item{x}{a digital elevation model raster layer}
}
\description{
Internal function for enerscape - calculate slope
}
