% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enerscape.R
\name{.calc_work}
\alias{.calc_work}
\title{Internal function for enerscape - calculate work}
\usage{
.calc_work(x, m, work_in_kcal = TRUE, j_to_kcal = 4184)
}
\arguments{
\item{x}{a slope transition layer}

\item{m}{species body mass (kg).}

\item{work_in_kcal}{if work should be expressed in kilocalories}

\item{j_to_kcal}{joules to kilocalories conversion constant}
}
\description{
Internal function for enerscape - calculate work
}
