% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/least-cost_path.R
\name{en_lcp}
\alias{en_lcp}
\title{Calculate the least-cost path (lcp) between origin and destination}
\usage{
en_lcp(en, neigh = 4, or, dest, simulate_random_points = FALSE, rep = 10)
}
\arguments{
\item{en}{an enerscape object obtained with enerscape::enerscape().}

\item{neigh}{number of neighbors used to calculate slope and work. See
?raster::terrain for details.}

\item{or}{origin point.}

\item{dest}{destination point.}

\item{simulate_random_points}{if to simulate least-cost path among random
points. default = FALSE.}

\item{rep}{number or random origin and destination points if
simulate_random_points = TRUE. default = 10.}
}
\description{
Calculate the least-cost path (lcp) between origin and destination
}
