% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enerscape.R
\name{enerscape}
\alias{enerscape}
\title{Calculate the energy landscape}
\usage{
enerscape(dem, m, units = "J")
}
\arguments{
\item{dem}{raster file of the digital elevation model, either a raster
or a full path location of the file}

\item{m}{is the body mass (kg) of the species}

\item{units}{if Joules ("J") or kilocalories ("kcal)}
}
\value{
raster of slope in degrees
}
\description{
Calculate the energy landscape
}
