% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enerscape_gridded.R
\name{enerscape_gridded}
\alias{enerscape_gridded}
\title{Run enerscape for a SpatialGrid}
\usage{
enerscape_gridded(
  dem,
  m,
  grid,
  unit = "joule",
  neigh = 16,
  method = "ARC",
  v = NULL,
  writedir = raster::tmpDir()
)
}
\arguments{
\item{dem}{raster file of the digital elevation model, either a raster or a
full path location of the file.}

\item{m}{species body mass (kg).}

\item{grid}{enerscape grid, as created using the \code{make_grid()} function.}

\item{unit}{if joules ('joule') or kilocalories ('kcal').}

\item{neigh}{number of neighbor cells that are connected together.}

\item{method}{method to use to compute the energy costs. 'ARC' refers to the
model from Pontzer (2016) and 'cycling' to the model for cyclist from di
Prampero et al. (1979).}

\item{v}{speed of cyclist (km / h), only for \code{method = 'cycling'}.}

\item{writedir}{directory to write EnergyScapes for each grid; default to tmpDir().}
}
\value{
A list with elements a rasterStack of the digital elevation model,
  slope, energy landscape, and conductance and the conductance as a transitionLayer for
  path analysis.
}
\description{
This runs enerscape for each block of a spatial grid sequentially.
}
\details{
\code{enerscape_gridded()} runs enerscape for each block of the grid
 sequentially and writes (only) the rasterLayer of the EnergyScape to the specified
 folder, default to the raster temporary directory. This temporary files are
 then merge back together by \code{enerscape_merge()}. This function also calculates
 enerscape for the vertical and horizontal strips used to split \code{dem} into blocks,
 which are then used to correct for artifacts of the splitting.
}
\examples{
\dontrun{
library(raster)
library(enerscape)
dem <- getData("alt", path = tmpDir(), country = "austria")
dem <- projectRaster(dem, crs = crs("EPSG:31255"), res = c(1000, 1000))
en_grid <- make_grid(extent(dem), crs(dem), 10, 10)
enerscape_gridded(dem, 10, grid = en_grid)
en <- enerscape_merge(tmpDir(), en_grid, dem)
}
}
\references{
Etten, J. van. (2017). R Package gdistance: Distances and Routes
  on Geographical Grids. Journal of Statistical Software, 76(1), 1–21.
  \doi{https://doi.org/10.18637/jss.v076.i13}.

  Pontzer, H. (2016). A unified theory for the energy cost of legged
  locomotion. Biology Letters, 12(2), 20150935. \doi{
  https://doi.org/10.1098/rsbl.2015.0935}.

  di Prampero, P. E., Cortili, G., Mognoni, P., & Saibene, F. (1979).
  Equation of motion of a cyclist. Journal of Applied Physiology, 47(1),
  201–206. \doi{https://doi.org/10.1152/jappl.1979.47.1.201}
}
