% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{slope}
\alias{slope}
\title{Slopes}
\usage{
slope(x, center, res)
}
\arguments{
\item{x}{matrix with values}

\item{center}{numeric value (double) with the value of the focal cell}

\item{res}{numeric value (double) of the spatial resolution of the matrix}
}
\value{
Vector with values the slopes (degrees) between x and center
}
\description{
Slopes
}
